AppDelegate.swift: no changes
MenuController.swift: no changes

GWNetworking.swift: Force unwrapped the title, issueNumber, volumeNumber and articleContent variables in the EditorialElement class (did not work, everything still has an Optional in front of it in the UI)

EditorialsTableViewController.swift: changed editorialObjects to a NSMutableOrderedSet with a capacity of 1000.
						      added everything required to format the timeStamp into a date object properly (timeStampDateString string, dateFormatter with a custom format of dd/MM/yy, timeStampDateObject, change to the prepareForSegue timeStamp variable --> timeStampDateString instead of timeStamp)
						      change to numberOfRowsInSection method --> editorialObjects.count instead of nodeIDArray.count
						      tried to add some force-unwrappings (!) to cellForRowAtIndexPath variables, but it did not work.
						      IMPORTANT: moved the "reloadData" function call to the outside of the "for node in JSON" loop since it made no sense to reload the data at every iteration of the for loop.

EditorialsTableViewCell.swift: Took out the editorialVolumeAndIssueLabel for the EditorialsTableViewController since it was taking too much space and was not that useful.

EditorialsDetailViewController.swift: changed the editorialDetailNavigationItem title from "\(editorialTitleThroughSegue)" to a static "Article" descriptor.

Main.storyboard: made a ton of little changes to the design to get the layout just right (not too much padding/margins, etc...).
			  took out the editorialVolumeAndIssueLabel label because it was unnecessary.