Minor code changes, major changes to the UI
CurrentIssueTableViewController.swift: minor changes

EditorialsTableViewController.swift: got rid of the Publish Date - the app will not be showing that in the detail view controller anymore.
EditorialsDetailViewController.swift: got rid of the Publish Date - the app will not be showing that in the detail view controller anymore.

NewsTableViewController.swift:  very minor change ("String(author) to author")
NewsDetailViewController.swift: got rid of the Publish Date - the app will not be showing that in the detail view controller anymore.

RandomTableViewController.swift: got rid of the Publish Date - the app will not be showing that in the detail view controller anymore.
RandomDetailViewController.swift: got rid of the Publish Date - the app will not be showing that in the detail view controller anymore.

Main.storyboard: major changes - adapted the user interface to all iPhone screen sizes (3.5-inch, 4-inch, 4.7-inch and 5.5-inch).
							--> reduced text size, modified constraints, deleted the Publish Date labels in the Detail View, changed some text colours.

Info.plist: disabled landscape mode.