WebView adjustment (--> disabled URL functionality), Map adjustment (subtitle), Assets
CurrentIssueTableViewController.swift: deleted a "else, return cell" statement. Not a major change.
CurrentIssueDetailViewController.swift: set the "dataDetectorType" to "None" for the webView, so that any link/address/other data that appeared in Articles would not be interpreted as a URL.

EditorialsDetailViewController.swift: set the "dataDetectorType" to "None" for the webView, so that any link/address/other data that appeared in Articles would not be interpreted as a URL.

NewsDetailViewController.swift: set the "dataDetectorType" to "None" for the webView, so that any link/address/other data that appeared in Articles would not be interpreted as a URL.

RandomDetailViewController.swift: set the "dataDetectorType" to "None" for the webView, so that any link/address/other data that appeared in Articles would not be interpreted as a URL.

IssueLocation.swift: changed the subtitle for all annotations/pins to "Queen's University", also changed the "CNPostalAddressStreetKey" name to "kABPersonAddressStreetKey" to make it compatible with iOS 8.0.

Assets.xcassets: Added a few assets, including the iTunesArtwork image set (which will be used as an app icon set for the App Store)

Main.storyboard: adjusted some constraints.

LaunchScreen.storyboard: no significant change (system-automated changes).

Info.plist: changed the Location Authorization Request message from "This will display your location on the campus map" to "This will allow you to get directions to the nearest Golden Words paper issue." (more precise).